refactor: AddBestAlbumTest 테스트 용도 변경
